# Optimize Sprintfln by using '\n' instead of fmt.Sprintln()