The benchmark result is:

benchmark               old ns/op     new ns/op     delta
BenchmarkSprintln       2939          2859          -2.72%
BenchmarkSprintln-2     2963          2904          -1.99%
BenchmarkSprintln-4     2994          2937          -1.90%
BenchmarkSprintln-8     3020          2920          -3.31%